+++ Makefile
@@ -1,4 +1,5 @@
 build:
+	go generate ./...
 	go build -o build/dp-frontend-renderer
 
 .PHONY: build
